requirements.txt
# Core
asyncio-compat>=0.1.0
panoramisk>=1.4

# ML
xgboost>=2.0.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Dashboard & Visualization
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
